+++ src/conn/dbs/postgres/model.mjs
@@ -1,9 +1,0 @@
-import CalustraModelBase from '../base/model.mjs'
-
-class CalustraModelPostgres extends CalustraModelBase {
-  constructor(connection, options) {
-    super(connection, options)
-  }
-}
-
-export default CalustraModelPostgres