+++ src/conn/index.mjs
@@ -1,6 +1,0 @@
-export { 
-  getConnection,
-  dropConnection,
-  dropConnections
-} from './api.mjs'
-