+++ test/query/index_dist.mjs
@@ -1,5 +1,5 @@
 
-import getTableNames from '../../src/query/getTableNames.mjs'
+import {getTableNames} from '../../dist/query/calustra-query.mjs'
 import test_table_names from './units/table_names.mjs'
 
 test_table_names(getTableNames)