+++ test/query/index_lib.mjs
@@ -1,0 +1,5 @@
+
+import getTableNames from '../../lib/query/getTableNames.cjs'
+import test_table_names from './units/table_names.mjs'
+
+test_table_names(getTableNames.default)